pyproject.toml
[project]
name = "slack-attendance-report"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.173.4",
    "constructs>=10.4.2",
    "slack-sdk>=3.34.0",
]

[tool.ruff.lint.isort]
force-single-line = true

[dependency-groups]
dev = [
    "boto3>=1.35.90",
    "faker>=33.1.0",
    "ipykernel>=6.29.5",
    "python-dotenv>=1.0.1",
]
